Moving all bot vote handling to vote.py
